Add support for privacy mask to blank out parts of image

+++ indi_allsky/detectLines.py
@@ -148,7 +148,7 @@ class IndiAllskyDetectLines(object):
             # color
             mask = cv2.cvtColor(blur_mask, cv2.COLOR_GRAY2BGR)
 
-        self._sqm_gradient_mask = mask / 255
+        self._sqm_gradient_mask = (mask / 255).astype(numpy.float16)
 
 
     def _drawLines(self, img, lines):
